# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Gin、Gorm 和 Quasar 的低代码管理底座",
        "title": "Sweet Admin",
        "contact": {
            "name": "南",
            "email": "maxdwy@gmail.com"
        },
        "version": "0.1"
    },
    "basePath": "/sweet_admin",
    "paths": {
...
        "/api/app_token": {
            "post": {
                "description": "获取AppToken",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "授权"
                ],
                "summary": "获取AppToken",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "自定义应用令牌",
                        "name": "X-APP-TOKEN",
                        "in": "header"
                    },
                    {
                        "description": "AppToken",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AppTokenReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/check_sms_status/{bizId}/{mobile}": {
            "get": {
                "description": "检查短信发送状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "授权"
                ],
                "summary": "检查短信发送状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "自定义应用令牌",
                        "name": "X-APP-TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "bizId",
                        "name": "bizId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "手机号",
                        "name": "mobile",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/login": {
            "post": {
                "description": "登录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "授权"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "自定义应用令牌",
                        "name": "X-APP-TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "登录信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SignInReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "description": "退出登录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "授权"
                ],
                "summary": "退出登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "自定义应用令牌",
                        "name": "X-APP-TOKEN",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/refresh_token": {
            "get": {
                "description": "刷新Token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "授权"
                ],
                "summary": "刷新Token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "刷新Token",
                        "name": "refreshToken",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/send_sms/{mobile}/{templateCode}": {
            "post": {
                "description": "发送短信",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "授权"
                ],
                "summary": "发送短信",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "自定义应用令牌",
                        "name": "X-APP-TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "发送短信",
                        "name": "data",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    {
                        "type": "string",
                        "description": "手机号",
                        "name": "mobile",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "短信模板编号",
                        "name": "templateCode",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/sms_code_login": {
            "post": {
                "description": "短信验证码登录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "授权"
                ],
                "summary": "短信验证码登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "自定义应用令牌",
                        "name": "X-APP-TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "短信验证码登录",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SmsCodeLoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/sso_login": {
            "get": {
                "description": "单点登录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "授权"
                ],
                "summary": "单点登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "自定义应用令牌",
                        "name": "X-APP-TOKEN",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.AppTokenReq": {
            "type": "object",
            "required": [
                "app_key",
                "app_secret"
            ],
            "properties": {
                "app_key": {
                    "type": "string"
                },
                "app_secret": {
                    "type": "string"
                }
            }
        },
...
        "request.SignInReq": {
            "type": "object",
            "required": [
                "password",
                "user_name"
            ],
            "properties": {
                "captcha": {
                    "type": "string"
                },
                "captcha_id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "request.SmsCodeLoginReq": {
            "type": "object",
            "required": [
                "code",
                "mobile"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}